+++ R/quo.R
@@ -162,7 +162,7 @@ as.quo.default <- function(x) {
 ifnot <- function(a, why) if (!isTRUE(a)) why else a
 `%&&%` <- function(a, b) if (!isTRUE(a)) a else b
 
-#' @export
+#' @exportS3Method all.equal quotation
 all.equal.quotation <- function(target, current, ...) {
   ifnot(is(current, "quotation"), "current is not a quotation") %&&%
     ifnot(all.equal(expr(target), expr(current), ...),