+++ src/vadr.h
@@ -5,6 +5,7 @@
 #include <Rinternals.h>
 
 // #undef DEBUG 
+// #define DEBUG
 
 #define MIN(x,y) ((x) < (y) ? (x) : (y))
 #define MAX(x,y) ((x) > (y) ? (x) : (y))